Flask/requirements.txt
# Core Framework
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9  # For PostgreSQL on Render

# ML/Data Handling
scikit-learn==1.5.0
pandas==2.2.2

# Utilities (based on common Flask setup, e.g., CORS)
Flask-CORS==4.0.1
python-dotenv==1.0.1